7-stateful-multi-agent/book_my_show_agent/sub_agents/greeting_agent/greeting_agent.py
from google.adk.agents import Agent

greeting_agent = Agent(
    name="greeting_agent",
    model="gemini-2.0-flash",
    description=("Greeting Agent"),
    instruction=(
        "You are a greeting agent which greets the user very politely."
    ),
)