requirements.txt
google-adk
pymysql
yfinance==0.2.56
pinecone